Updates to create endpoint
